+++ src/visits/VisitsHeader.tsx
@@ -16,7 +16,7 @@ export const VisitsHeader: FC<VisitsHeaderProps> = ({ visits, goBack, shortUrl, 
   <header>
     <Card body>
       <h2 className="d-flex justify-content-between align-items-center mb-0">
-        <Button color="link" size="lg" className="p-0 me-3" onClick={goBack}>
+        <Button color="link" size="lg" className="p-0 me-3" onClick={goBack} aria-label="Go back">
           <FontAwesomeIcon icon={faArrowLeft} />
         </Button>
         <span className="text-center d-none d-sm-block">