+++ test/visits/NonOrphanVisits.test.tsx
@@ -6,6 +6,7 @@ import type { MercureBoundProps } from '../../src/mercure/helpers/boundToMercure
 import { SettingsProvider } from '../../src/utils/settings';
 import { NonOrphanVisitsFactory } from '../../src/visits/NonOrphanVisits';
 import type { VisitsInfo } from '../../src/visits/reducers/types';
+import { checkAccessibility } from '../__helpers__/accessibility';
 import { renderWithEvents } from '../__helpers__/setUpTest';
 
 describe('<NonOrphanVisits />', () => {
@@ -28,6 +29,8 @@ describe('<NonOrphanVisits />', () => {
       </SettingsProvider>
     </MemoryRouter>,
   );
+
+  it('passes a11y checks', () => checkAccessibility(setUp()));
 
   it('wraps visits stats and header', () => {
     setUp();