+++ test/visits/TagVisitsHeader.test.tsx
@@ -3,6 +3,7 @@ import { fromPartial } from '@total-typescript/shoehorn';
 import type { ColorGenerator } from '../../src/utils/services/ColorGenerator';
 import type { TagVisits } from '../../src/visits/reducers/tagVisits';
 import { TagVisitsHeader } from '../../src/visits/TagVisitsHeader';
+import { checkAccessibility } from '../__helpers__/accessibility';
 
 describe('<TagVisitsHeader />', () => {
   const tagVisits = fromPartial<TagVisits>({
@@ -12,6 +13,8 @@ describe('<TagVisitsHeader />', () => {
   const goBack = vi.fn();
   const colorGenerator = fromPartial<ColorGenerator>({ isColorLightForKey: () => false, getColorForKey: () => 'red' });
   const setUp = () => render(<TagVisitsHeader tagVisits={tagVisits} goBack={goBack} colorGenerator={colorGenerator} />);
+
+  it('passes a11y checks', () => checkAccessibility(setUp()));
 
   it('shows expected visits', () => {
     const { container } = setUp();